Automatically discover unit tests
Use a simple of convention and a bit of code generation
to automatically populate the list of test suites.
